pyproject.toml
[project]
name = "white_elephant_bot"
version = "0.1.0"
description = "Bot for the Whalefall Discord server"
authors = [
    {name = "Matt Zhang", email = "set.stun@gmail.com"},
]
dependencies = [
    "slash-discord-py",
    "python-dotenv",
    "fastapi",
    "hypercorn",
    "pynacl",
    "discord>=2.3.2",
    "py-gpt-interface>=0.6.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
